fix(F-83): rename props file
