+++ frontend/src/interfaces/StepperProps.ts
