+++ frontend/src/styles/globals.css
@@ -41,7 +41,6 @@
   /* typography */
   --text-2xl: 24px;
   --text-2xl--font-weight: 700;
-  --text-icon: 20px;
   --text-xl: 18px;
   --text-xl--font-weight: 600;
   --text-lg: 13px;
